requirements.txt
flask
opencv-python
numpy
scikit-learn
Pillow
joblib